Added characteristics selection into new component
